move DummyPacket to api package
